backend/requirements.txt
appdirs==1.4.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==7.1.2
cloudscraper==1.2.71
colorama==0.4.6
idna==3.10
pixiv-api==0.3.7
pybit==5.11.0
pycryptodome==3.23.0
pyparsing==3.2.3
requests==2.32.4
requests-toolbelt==1.0.0
tqdm==4.67.1
urllib3==2.5.0
websocket-client==1.8.0
djangorestframework-simplejwt 